+++ src/EVM/UnitTest.hs
@@ -13,17 +13,18 @@ import EVM.FeeSchedule (feeSchedule)
 import EVM.Fetch qualified as Fetch
 import EVM.Format
 import EVM.Solidity
-import EVM.SymExec (defaultVeriOpts, symCalldata, verify, extractCex, prettyCalldata, panicMsg, VeriOpts(..), flattenExpr, groupIssues, groupPartials, IterConfig(..), defaultIterConf, LoopHeuristic)
+import EVM.SymExec (defaultVeriOpts, symCalldata, verify, extractCex, prettyCalldata, calldataFromCex, panicMsg, VeriOpts(..), flattenExpr, groupIssues, groupPartials, IterConfig(..), defaultIterConf, LoopHeuristic)
 import EVM.Types
 import EVM.Transaction (initTx)
 import EVM.Stepper (Stepper)
 import EVM.Stepper qualified as Stepper
+import EVM.Tracing qualified as Tracing
 import EVM.Expr (maybeLitWordSimp)
 import Data.List (foldl')
 
 import Control.Monad (void, when, forM, forM_)
 import Control.Monad.ST (RealWorld, ST, stToIO)
-import Control.Monad.State.Strict (execState, get, put, liftIO)
+import Control.Monad.State.Strict (execState, get, put, liftIO, runStateT)
 import Optics.Core
 import Optics.State
 import Optics.State.Operators
@@ -160,6 +161,42 @@ initializeUnitTest opts theContract = do
     Left e -> pushTrace (ErrorTrace e)
     _ -> popTrace
 
+validateCex :: App m
+  => UnitTestOptions RealWorld
+  -> VM Concrete RealWorld
+  -> ReproducibleCex
+  -> m Bool
+validateCex uTestOpts vm repCex = do
+  let utoConc = uTestOpts { testParams = uTestOpts.testParams { caller = LitAddr 0xdeadbeef}}
+  conf <- readConfig
+  when conf.debug $ liftIO $ putStrLn $ "Repro running function: " <> show utoConc.testParams.address <>
+    " with caller: " <> show utoConc.testParams.caller <> ", gas: " <> show utoConc.testParams.gasCall <>
+    ", and calldata: " <> show (bsToHex repCex.callData)
+
+  -- Note, we should not need solvers to execute this code
+  vm2 <- Stepper.interpret (Fetch.oracle utoConc.solvers utoConc.rpcInfo) vm $ do
+    Stepper.evm $ do
+      pushTrace (EntryTrace $ "checking cex for function " <> repCex.testName <> " with calldata: " <> (Text.pack $ bsToHex repCex.callData))
+      makeTxCall utoConc.testParams (ConcreteBuf repCex.callData, [])
+    Stepper.evm get
+
+  (res, (vm3, vmtrace)) <- runStateT (Tracing.interpretWithTrace (Fetch.oracle utoConc.solvers utoConc.rpcInfo) Stepper.execFully) (vm2, [])
+  when conf.debug $ liftIO $ do
+    putStrLn $ "vm step trace: " <> unlines (map show vmtrace)
+    putStrLn $ "vm res: " <> show res
+    putStrLn $ "vm res: " <> show vm3.result
+  let shouldFail = "proveFail" `isPrefixOf` repCex.testName
+  let ret = case res of
+              Left err -> case err of
+                (UnrecognizedOpcode 0xfe) -> True
+                (Revert (ConcreteBuf msg)) ->
+                   msg == panicMsg 0x01 ||
+                     let sel = selector "Error(string)"
+                     in (sel `BS.isPrefixOf` msg) && ("assertion failed" `BS.isPrefixOf` (BS.drop (4+32+32) msg))
+                _ -> False
+              Right _ -> utoConc.checkFailBit && (dsTestFailedConc vm3.env.contracts)
+  pure (ret /= shouldFail)
+
 -- Returns tuple of (No Cex, No warnings)
 runUnitTestContract
   :: App m
@@ -193,26 +230,35 @@ runUnitTestContract
           forM testSigs $ \s -> symRun opts vm1 s solcContr buildOut.sources
         _ -> internalError "setUp() did not end with a result"
 
+dsTestFailedSym :: Map (Expr 'EAddr) (Expr EContract) -> VM s t -> Prop
+dsTestFailedSym store vm =
+  let testContract = fromMaybe (internalError "test contract not found in state") (Map.lookup vm.state.contract store)
+  in case Map.lookup cheatCode store of
+    Just cheatContract -> Expr.readStorage' (Lit 0x6661696c65640000000000000000000000000000000000000000000000000000) cheatContract.storage .== Lit 1
+    Nothing -> And (Expr.readStorage' (Lit 0) testContract.storage) (Lit 2) .== Lit 2
+
+dsTestFailedConc :: Map (Expr 'EAddr) Contract -> Bool
+dsTestFailedConc store = case Map.lookup cheatCode store of
+  Just cheatContract -> Expr.readStorage' (Lit 0x6661696c65640000000000000000000000000000000000000000000000000000) cheatContract.storage == Lit 1
+  Nothing -> internalError "dsTestFailedConc: expected a cheatCode in the store"
+
 -- Define the thread spawner for symbolic tests
 -- Returns tuple of (No Cex, No warnings)
 symRun :: App m => UnitTestOptions RealWorld -> VM Concrete RealWorld -> Sig -> SolcContract -> SourceCache -> m (Bool, Bool)
-symRun opts@UnitTestOptions{..} vm (Sig testName types) solcContr sourceCache = do
-    let callSig = testName <> "(" <> (Text.intercalate "," (map abiTypeSolidity types)) <> ")"
-    liftIO $ putStrLn $ "\x1b[96m[RUNNING]\x1b[0m " <> Text.unpack callSig
-    cd <- symCalldata callSig types [] (AbstractBuf "txdata")
-    let shouldFail = "proveFail" `isPrefixOf` callSig
+symRun opts@UnitTestOptions{..} vm sig@(Sig testName types) solcContr sourceCache = do
+    let cs = callSig sig
+    liftIO $ putStrLn $ "\x1b[96m[RUNNING]\x1b[0m " <> Text.unpack cs
+    cd <- symCalldata cs types [] (AbstractBuf "txdata")
+    let shouldFail = "proveFail" `isPrefixOf` cs
 
     -- define postcondition depending on `shouldFail`
-    let testContract store = fromMaybe (internalError "test contract not found in state") (Map.lookup vm.state.contract store)
-        failed store = case Map.lookup cheatCode store of
-          Just cheatContract -> Expr.readStorage' (Lit 0x6661696c65640000000000000000000000000000000000000000000000000000) cheatContract.storage .== Lit 1
-          Nothing -> And (Expr.readStorage' (Lit 0) (testContract store).storage) (Lit 2) .== Lit 2
+    let
         postcondition = curry $ case shouldFail of
           True -> \(_, post) -> case post of
-            Success _ _ _ store -> if opts.checkFailBit then failed store else PBool False
+            Success _ _ _ store -> if opts.checkFailBit then (dsTestFailedSym store vm) else PBool False
             _ -> PBool True
           False -> \(_, post) -> case post of
-            Success _ _ _ store -> if opts.checkFailBit then PNeg (failed store) else PBool True
+            Success _ _ _ store -> if opts.checkFailBit then PNeg (dsTestFailedSym store vm) else PBool True
             Failure _ _ (UnrecognizedOpcode 0xfe) -> PBool False
             Failure _ _ (Revert msg) -> case msg of
               ConcreteBuf b ->
@@ -242,7 +288,6 @@ symRun opts@UnitTestOptions{..} vm (Sig testName types) solcContr sourceCache = 
       _ -> internalError "cannot be, filtered for failure"
 
     -- display results
-    let t = "the test " <> Text.unpack testName
     let warnings = any Expr.isPartial ends || any isUnknown results || any isError results
     let allReverts = not . (any Expr.isSuccess) $ ends
     let unexpectedAllRevert = allReverts && not shouldFail
@@ -254,8 +299,12 @@ symRun opts@UnitTestOptions{..} vm (Sig testName types) solcContr sourceCache = 
       (True, _, _) -> do
         -- there are counterexamples (and maybe other things, but Cex is most important)
         let x = mapMaybe extractCex results
-        y <- symFailure opts testName (fst cd) types x
-        pure $ "   \x1b[31m[FAIL]\x1b[0m " <> Text.unpack testName <> "\n" <> Text.unpack y
+        failsToRepro <- getReproFailures (Sig testName types) (fst cd) (map snd x)
+        validation <- mapM (traverse $ validateCex opts vm) failsToRepro
+        when conf.debug $ liftIO $ putStrLn $ "Cex reproduction runs' results are: " <> show validation
+        let toPrintData = zipWith (\(a, b) c -> (a, b, c)) x validation
+        txtFails <- symFailure opts testName (fst cd) types toPrintData
+        pure $ "   \x1b[31m[FAIL]\x1b[0m " <> Text.unpack testName <> "\n" <> Text.unpack txtFails
       (_, True, _) -> do
         -- There are errors/unknowns/partials, we fail them
         pure $ "   \x1b[31m[FAIL]\x1b[0m " <> Text.unpack testName <> "\n"
@@ -268,7 +317,7 @@ symRun opts@UnitTestOptions{..} vm (Sig testName types) solcContr sourceCache = 
       -- if we display a FAILED due to Cex/warnings, we should also mention everything reverted
       liftIO $ putStrLn $ "   \x1b[33m[WARNING]\x1b[0m " <> Text.unpack testName <> " all branches reverted\n"
     let warnData = Just $ WarningData solcContr sourceCache vm'
-    liftIO $ printWarnings warnData ends results t
+    liftIO $ printWarnings warnData ends results $ "the test " <> Text.unpack testName
     pure (not (any isCex results), not (warnings || unexpectedAllRevert))
     where
       -- The offset of the text is: the selector (4B), the offset value (aligned to 32B), and the length of the string (aligned to 32B)
@@ -305,10 +354,20 @@ printWarnings warnData e results testName = do
     forM_ (groupPartials warnData e) $ \(num, str) -> putStrLn $ "      " <> show num <> "x -> " <> str
   putStrLn ""
 
-symFailure :: App m => UnitTestOptions RealWorld -> Text -> Expr Buf -> [AbiType] -> [(Expr End, SMTCex)] -> m Text
-symFailure UnitTestOptions {..} testName cd types failures' = do
+getReproFailures :: App m => Sig -> Expr Buf -> [SMTCex] -> m [Err ReproducibleCex]
+getReproFailures sig@(Sig testName _) cd cexes = do
+  fullCDs <- mapM (\cex -> calldataFromCex cex cd sig) cexes
+  pure $ map (\case
+    Left err -> Left err
+    Right fullCD -> Right $ ReproducibleCex { testName = testName, callData = fullCD}) fullCDs
+
+symFailure :: App m =>
+  UnitTestOptions RealWorld -> Text -> Expr Buf -> [AbiType] ->
+  [(Expr End, SMTCex, Err Bool)] ->
+  m Text
+symFailure UnitTestOptions {..} testName cd types fails = do
   conf <- readConfig
-  pure $ mconcat [ Text.concat $ indentLines 3 . mkMsg conf <$> failures' ]
+  pure $ mconcat [ Text.concat $ indentLines 3 . mkMsg conf <$> fails ]
   where
       showRes = \case
         Success _ _ _ _ -> if "proveFail" `isPrefixOf` testName
@@ -317,8 +376,8 @@ symFailure UnitTestOptions {..} testName cd types failures' = do
         res ->
           let ?context = dappContext (traceContext res)
           in Text.pack $ prettyvmresult res
-      mkMsg conf (leaf, cex) = intercalate "\n" $
-        ["Counterexample:"
+      mkMsg conf (leaf, cex, repro) = intercalate "\n" $
+        ["Counterexample:   " <> reproToText repro
         ,"  calldata: " <> let ?context = dappContext (traceContext leaf)
                            in prettyCalldata cex cd testName types
         ,"  result:   " <> showRes leaf
@@ -327,6 +386,10 @@ symFailure UnitTestOptions {..} testName cd types failures' = do
                            else [Text.unlines [ indentLines 2 (showTraceTree' dapp leaf)]]
       dappContext TraceContext { contracts, labels } =
         DappContext { info = dapp, contracts, labels }
+
+      reproToText :: Err Bool -> Text
+      reproToText (Left err) = "\x1b[33m[error attempting to reproduce: " <> pack err <> "]\x1b[0m"
+      reproToText (Right repro) = if repro then "\x1b[90m[validated]\x1b[0m" else "\x1b[31m[not reproducible]\x1b[0m"
 
 indentLines :: Int -> Text -> Text
 indentLines n s =